+++ cmd/pr_has_fragment.go
@@ -20,8 +20,8 @@ var errPrCheckCmdMissingArg = errors.New("pr-has-fragment command requires pr nu
 
 func PrHasFragmentCommand(appFs afero.Fs) *cobra.Command {
 	prCheckCmd := &cobra.Command{
-		Use:  "pr-has-fragment <pr-number>",
-		Long: "Check changelog fragment presence in the given PR.",
+		Use:   "pr-has-fragment <pr-number>",
+		Short: "Check changelog fragment presence in the given PR.",
 		Args: func(cmd *cobra.Command, args []string) error {
 			if len(args) < 1 {
 				return errPrCheckCmdMissingArg
